chapter 6 & 7: setting db on vercel and how to fetch data in nextjs! so in nextjs, because the react components are rendered on the server, we can directly fetch data on the server! and directly communicate to the db without the need of an API layer! noice!
we can make our react components to be async and just await the data directly inside the react component!
the documentation also suggest to always fetch specific data that is needed.
for e.g. don't query the whole data to get the count of the table rows. instead use SQL or an ORM of your choice to get the desired information about your tables! it will translate to better performance gains because the information is handled with the query, not by javascript.

but, there are 2 things to be aware of:
-using await on one to the other network request on the same ccomponent can lead to request waterfall.
- nextjs by default, prerenders routes to improve performance (static rendering). so if at one point, your data changes, it won't be reflected in your rendered dashboard.

we can negate the request waterfall by fetchin data sequentially, or by parallel. in JS, we can use Promise.all() or Promise.allSettled() funcs to initiate all promises at the same time!