chapter 8: static and dynamic fetching! so we can use static or dynamic fethcing in nextjs static fetching is good when you are showing the users that are the same for everyone and is not constantly changes, e.g. blogs, products and what not. but it has a downside of not reflecting updated changes to the users. because static fetching is fetched during the build time and is cached on the server! well, we can negate that with unvalidating the cahce by time or by an event. but we haven't reached to that point yet :)
for preventing stale data, we can make the use of dynamic fetching!
dynamic fetching is good for the use of frequently changing data. for example, for a dashboard, we always need to show the users the most updated version of the data. so we ccan use dynamic fetching by fetching the data and with the use of unstable_noStore() from next/cache. it just helps to tell nextjs, to not cache our data!

but the downside is we lost the benefits of using static rendering. that is:
-faster websites, cuz our web's components are prerendered on the server
-reduced server load, cuz the content is cached, our server doesn't have to constantly generate content for each user request, just give 'em the prerendered page! haha
-good SEO, prerendered content makes search engine bots/crawlers happy! better indexes, improve search engine rankings!
good stuff indeed.

one quote from the documentation that I like,
'With dynamic rendering, your application is only as fast as your slowest data fetch.'