chapter 13: Handling Errors we can handle errros pretty neatly in nextjs! with the help of trycatch block ofc! simply in our async functions, there are some promise functions, which can fail, so we wrap em up with trycatch, in the catch block, we can just throw new Error(errMsg). if the fucntion results on an error, it will go to the catch block and throws an error. now how do we handle it so that our app is still usable and we can let the users know that there is something wrong?!
we can do it by creating an error.tsx file onour root file. it basically just a catch all error. a fallback page if u may. we can get the error from the special prop that error.tsx can get, which are error and reset.

error is the error instance itself, and reset is a method to rerender the page.

but what about 404 page? we can just create not-found.tsx file. it's sole purpose is to handle not found route. good stuff indeeed.