chapter 14: improve accessibility, useFormStatus, and server side valdiation! to increase accessibility to our form, there are several things to keep in mind: - semantic HTML, labeling and attaching htmlFor attr, and Focux outline!
for form validation, we can do it in two ways, either client-side valdiation, or server side valdiation.
width client side validation, we can implement the use of required attribute on the input tag, this will trigger validation on the browser lebel and u can get the browser warning. it's an okay valdiation, but there are alternative way with the use of server side valdiation.

with server side validation, you get several benefits:
- ensure the data received is in the expected format before sending it to db, reduce the risk of malicious users bypassing cliendside validation, and have one source of truth for what is considered valid data!

to make use of server side validation, we can do it with the help of useFormState hook, it is a hook, so ur form component needs to be a client side component. dont fotget to use the 'use client' directive on the top of ur form component file.
then, just call the useFormState, it takes 2 arguments, action and initial state. and will return two values in an array, [state, dispatch] that is. similar to useReducer if u may.

then, we can pass our server action to first argument of the hook, and don't forget to create an initial state {message:null, errors: {}} to be passed as the second argument. the initial state is an object like that cuz we're just taking the consideration of what zod returns.
then just pass the dispatch to the form action. and we can access the errors message from the state returned from the hook.