chapter 16: added metadatas! metadata is crucial for SEO and shareability! some important metadatas: title, description, keywords, opengraphs, and favicon.
in nextjs, we can include a metadata object on our layout.ts to serve as the metadata for all pages. the metadata in the layout will be inheriter by all pages that uses it. But what if u want to add the title's metadata based on each page? can u just add metadata object on every page? well yes, but there's a way to do that.
we can use the title.template field in th emetadata object on ur layout page. like this: title: {template: ' | example', default: 'example'}. the '%s' in the template will be replaced with the specific page title.

good stuff